# Conflicts:
#	Gainy/Gainy.xcodeproj/project.pbxproj
#	Gainy/Gainy/Models/RemoteTicker.swift
#	Gainy/Gainy/Scenes/Discovery/TickerDetails/Controller/Cells/Charts/ScatterChartView.swift
#	Gainy/Gainy/Scenes/Home/HomeViewController.swift
#	Gainy/Gainy/Scenes/Home/HomeViewModel.swift
#	Gainy/Gainy/Scenes/Portfolio/Plaid/Cells/Holdings/HoldingSecurityViewModel.swift
#	Gainy/Gainy/Scenes/Portfolio/Plaid/Cells/Holdings/HoldingTableViewCell.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/AppInterests.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/Categories.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/Charts.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/CollectionDetails.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchLiveTickersData.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchRealtimeMetrics.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchSelectedCollections.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchStockMedian.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchStockMedianMonth.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchStockMedianWeek.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/FetchTickers.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/GetNextTickersForCollection.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/GetPlaidHoldings.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/GetSelectedCategories.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/GetSelectedInterests.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/GetTickerExtraDetails.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/HomeFetchArticles.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/HomeFetchGainers.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/SearchCollectionDetails.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/SearchTickers.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/SearchTickersNoSpace.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/SearchTickersWithSpace.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/API/Types.graphql.swift
#	Gainy/Gainy/Services/GraphQL/Schema/schema.json
#	ReleaseNotes.txt

+++ Gainy/Gainy/Services/GraphQL/Schema/API/FetchStockMedian.graphql.swift
@@ -8,7 +8,7 @@ public final class FetchStockMedianQuery: GraphQLQuery {
   /// The raw GraphQL definition of this operation.
   public let operationDefinition: String =
     """
-    query FetchStockMedian($industryId: bigint!, $period: String!) {
+    query FetchStockMedian($industryId: Int!, $period: String!) {
       industry_median_chart(
         where: {industry_id: {_eq: $industryId}, period: {_eq: $period}}
       ) {
@@ -53,7 +53,7 @@ public final class FetchStockMedianQuery: GraphQLQuery {
       self.init(unsafeResultMap: ["__typename": "query_root", "industry_median_chart": industryMedianChart.map { (value: IndustryMedianChart) -> ResultMap in value.resultMap }])
     }
 
-    /// fetch data from the table: "public_220413044555.industry_median_chart"
+    /// fetch data from the table: "public_220413060411.industry_median_chart"
     public var industryMedianChart: [IndustryMedianChart] {
       get {
         return (resultMap["industry_median_chart"] as! [ResultMap]).map { (value: ResultMap) -> IndustryMedianChart in IndustryMedianChart(unsafeResultMap: value) }
